initial refactor
Refactore in more structured way. Graph structure without member functions. It's only first step of refactoring. The next step should be to refactore piles
