+++ src/Graph/GraphAssemble.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "./Graph/Graph.hpp"
+
+namespace raven {
+
+// Layout phase
+// - remove transitive edges
+// - remove tips
+// - remove bubbles
+// - remove elongated edges in 2D layout
+void assemble(std::shared_ptr<thread_pool::ThreadPool> threadPool, Graph& graph, bool checkpoints);
+
+}
+
