+++ src/Graph/Serialization/GraphBinarySerialization.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+
+#include "Graph/Graph.hpp"
+
+namespace raven {
+
+void storeGraphToFile(const Graph& graph);
+Graph loadGraphFromFile();
+
+}
